Fixed build errors on Visual C++ 2010.
git-svn-id: https://xemmai.googlecode.com/svn/trunk@134 21b6bbce-954d-0410-9621-597f947845d9

+++ include/xemmai/portable/atomic.h
@@ -130,12 +130,12 @@ public:
 #ifdef _MSC_VER
 inline size_t f_atomic_increment(volatile size_t& a_n)
 {
-	return _InterlockedIncrement(reinterpret_cast<volatile ptrdiff_t*>(&a_n));
+	return _InterlockedIncrement(reinterpret_cast<volatile long*>(&a_n));
 }
 
 class t_lock
 {
-	volatile size_t v_lock;
+	volatile long v_lock;
 
 public:
 	t_lock() : v_lock(0)
